+++ apps/wsb-task/src/commands/mod.rs
@@ -1,5 +1,3 @@
 mod artifacts;
 pub use artifacts::copy_as_artifact;
 pub use artifacts::{app_names, artifacts_dir, executable_names};
-
-pub mod support;
