+++ apps/wsb-task/src/core/mod.rs
@@ -1,5 +1,7 @@
 mod action;
 pub use action::ActionOutput;
 
+pub mod support;
+
 #[macro_use]
 pub mod targets;
