+++ apps/wsb-task/src/tasks/assemble_pilot_tests/task/build_pilot/params.rs
@@ -1,4 +1,4 @@
-use crate::commands::support::CCFindable;
+use crate::core::support::CCFindable;
 use crate::core::targets::{AsBuildTarget, BuildTarget};
 use crate::core::ActionOutput;
 use std::path::PathBuf;
