+++ apps/wsb-task/src/tasks/build_apps/task.rs
@@ -1,4 +1,4 @@
-use crate::commands::support::CCFindable;
+use crate::core::support::CCFindable;
 use crate::TaskResult;
 use shellwork::core::command;
 use shellwork::core::command::{no_op, Prepared, Runner, Unprepared};
