+++ apps/wsb-task/src/tasks/shared/params.rs
@@ -1,4 +1,4 @@
-use crate::commands::support::CCFindable;
+use crate::core::support::CCFindable;
 use crate::core::targets::{AsBuildTarget, BuildTarget};
 use crate::tasks::shared::build_target;
 use crate::TaskResult;
